metadata_creation/pyproject.toml
[project]
name = "metadata-creation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "neo4j>=5.28.1",
    "pinecone>=8.1.1",
    "python-dotenv>=1.2.2",
    "sqlitecloud>=0.0.84",
]
